+++ tests/snowpark/test_procedure.py
@@ -53,26 +53,26 @@ def test_deploy_procedure(
         ]
     )
     assert ctx.get_queries() == [
-        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT comment='deployments managed by Snowflake CLI'",
-        f"put file://{Path(tmp).resolve()}/app.zip @MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT/my_snowpark_project auto_compress=false parallel=4 overwrite=True",
+        "create stage if not exists MockDatabase.MockSchema.dev_deployment comment='deployments managed by Snowflake CLI'",
+        f"put file://{Path(tmp).resolve()}/app.zip @MockDatabase.MockSchema.dev_deployment/my_snowpark_project auto_compress=false parallel=4 overwrite=True",
         dedent(
             """\
-            create or replace procedure MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(name string)
+            create or replace procedure MockDatabase.MockSchema.procedureName(name string)
             returns string
             language python
             runtime_version=3.8
-            imports=('@MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT/my_snowpark_project/app.zip')
+            imports=('@MockDatabase.MockSchema.dev_deployment/my_snowpark_project/app.zip')
             handler='hello'
             packages=()
             """
         ).strip(),
         dedent(
             """\
-            create or replace procedure MOCKDATABASE.MOCKSCHEMA.TEST()
+            create or replace procedure MockDatabase.MockSchema.test()
             returns string
             language python
             runtime_version=3.10
-            imports=('@MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT/my_snowpark_project/app.zip')
+            imports=('@MockDatabase.MockSchema.dev_deployment/my_snowpark_project/app.zip')
             handler='test'
             packages=()
             """
@@ -113,21 +113,21 @@ def test_deploy_procedure_with_external_access(
         [
             call(
                 object_type=str(ObjectType.PROCEDURE),
-                name="MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(string)",
+                name="MockDatabase.MockSchema.procedureName(string)",
             ),
         ]
     )
     assert ctx.get_queries() == [
-        "create stage if not exists MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT comment='deployments managed by Snowflake CLI'",
-        f"put file://{Path(project_dir).resolve()}/app.zip @MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT/my_snowpark_project"
+        "create stage if not exists MockDatabase.MockSchema.dev_deployment comment='deployments managed by Snowflake CLI'",
+        f"put file://{Path(project_dir).resolve()}/app.zip @MockDatabase.MockSchema.dev_deployment/my_snowpark_project"
         f" auto_compress=false parallel=4 overwrite=True",
         dedent(
             """\
-            create or replace procedure MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(name string)
+            create or replace procedure MockDatabase.MockSchema.procedureName(name string)
             returns string
             language python
             runtime_version=3.8
-            imports=('@MOCKDATABASE.MOCKSCHEMA.DEV_DEPLOYMENT/my_snowpark_project/app.zip')
+            imports=('@MockDatabase.MockSchema.dev_deployment/my_snowpark_project/app.zip')
             handler='app.hello'
             packages=()
             external_access_integrations=(external_1,external_2)
@@ -255,7 +255,7 @@ def test_deploy_procedure_replace_nothing_to_update(
                 ("packages", "[]"),
                 ("handler", "hello"),
                 ("returns", "string"),
-                ("imports", "DEV_DEPLOYMENT/my_snowpark_project/app.zip"),
+                ("imports", "dev_deployment/my_snowpark_project/app.zip"),
             ],
             columns=["key", "value"],
         ),
@@ -264,7 +264,7 @@ def test_deploy_procedure_replace_nothing_to_update(
                 ("packages", "[]"),
                 ("handler", "test"),
                 ("returns", "string"),
-                ("imports", "DEV_DEPLOYMENT/my_snowpark_project/app.zip"),
+                ("imports", "dev_deployment/my_snowpark_project/app.zip"),
             ],
             columns=["key", "value"],
         ),
@@ -278,12 +278,12 @@ def test_deploy_procedure_replace_nothing_to_update(
     assert result.exit_code == 0, result.output
     assert json.loads(result.output) == [
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(name string)",
+            "object": "MockDatabase.MockSchema.procedureName(name string)",
             "status": "packages updated",
             "type": "procedure",
         },
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.TEST()",
+            "object": "MockDatabase.MockSchema.test()",
             "status": "packages updated",
             "type": "procedure",
         },
@@ -308,7 +308,7 @@ def test_deploy_procedure_replace_updates_single_object(
                 ("packages", "[]"),
                 ("handler", "hello"),
                 ("returns", "string"),
-                ("imports", "DEV_DEPLOYMENT/my_snowpark_project/app.zip"),
+                ("imports", "dev_deployment/my_snowpark_project/app.zip"),
             ],
             columns=["key", "value"],
         ),
@@ -317,7 +317,7 @@ def test_deploy_procedure_replace_updates_single_object(
                 ("packages", "[]"),
                 ("handler", "foo"),
                 ("returns", "string"),
-                ("imports", "DEV_DEPLOYMENT/my_snowpark_project/app.zip"),
+                ("imports", "dev_deployment/my_snowpark_project/app.zip"),
             ],
             columns=["key", "value"],
         ),
@@ -331,12 +331,12 @@ def test_deploy_procedure_replace_updates_single_object(
     assert result.exit_code == 0
     assert json.loads(result.output) == [
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(name string)",
+            "object": "MockDatabase.MockSchema.procedureName(name string)",
             "status": "packages updated",
             "type": "procedure",
         },
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.TEST()",
+            "object": "MockDatabase.MockSchema.test()",
             "status": "definition updated",
             "type": "procedure",
         },
@@ -361,7 +361,7 @@ def test_deploy_procedure_replace_creates_missing_object(
                 ("packages", "[]"),
                 ("handler", "hello"),
                 ("returns", "string"),
-                ("imports", "DEV_DEPLOYMENT/my_snowpark_project/app.zip"),
+                ("imports", "dev_deployment/my_snowpark_project/app.zip"),
             ],
             columns=["key", "value"],
         ),
@@ -376,12 +376,12 @@ def test_deploy_procedure_replace_creates_missing_object(
     assert result.exit_code == 0
     assert json.loads(result.output) == [
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.PROCEDURENAME(name string)",
+            "object": "MockDatabase.MockSchema.procedureName(name string)",
             "status": "packages updated",
             "type": "procedure",
         },
         {
-            "object": "MOCKDATABASE.MOCKSCHEMA.TEST()",
+            "object": "MockDatabase.MockSchema.test()",
             "status": "created",
             "type": "procedure",
         },